make the code modular
