+++ component/scripts/__init__.py
@@ -1,3 +1,5 @@
 from .export import *
 from .table import *
 from .vizualisation import *
+from .gee import *
+from .planet import *
